# backend/docs/swagger.yaml
basePath: /api
host: edis.dshub.com.ng
info:
  contact:
    email: support@lifegate.health
    name: LifeGate Support
  description: AI-powered clinical diagnostic platform. Patients interact with an
    AI assistant, physicians review AI-generated reports, and administrators manage
    the full lifecycle.
  license:
    name: Proprietary
  termsOfService: https://lifegate.health/terms
  title: LifeGate API
  version: "1.0"
paths:
...
  /review/analysis:
    get:
      description: Returns daily diagnosis counts for the given date range (default
        last 30 days) plus 10 recent activities.
      parameters:
      - description: Single date (YYYY-MM-DD). Overrides startDate/endDate.
        in: query
        name: date
        type: string
      - description: Start of range (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End of range (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              activities:
                type: array
              data:
                type: array
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Physician analysis
      tags:
      - review
  /review/diagnoses:
    get:
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: pageSize
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Start date filter (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: End date filter (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  records:
                    type: array
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: List physician diagnoses
      tags:
      - review
  /review/diagnoses/{id}:
    get:
      parameters:
      - description: Diagnosis ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Get review diagnosis detail
      tags:
      - review
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"